aws-python/client/requirements.txt
paho-mqtt
matplotlib
pandas
keyboard
